requirements.txt
neat-python
pygame
pygame-gui
matplotlib